# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sabah Park Api Documentation",
        "contact": {
            "email": "prastiyo.beka12@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/dashboard/summary": {
            "get": {
                "tags": [
                    "DASHBOARD"
                ],
                "summary": "Get Data Member",
                "operationId": "App\\Http\\Controllers\\Admin\\DashboardController::summary",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/admin/ticket/get": {
            "get": {
                "tags": [
                    "Setup Ticket (BackOffice)"
                ],
                "summary": "Get Data Ticket",
                "operationId": "App\\Http\\Controllers\\Admin\\TicketController::get",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/admin/ticket/scan": {
            "get": {
                "tags": [
                    "Setup Ticket (BackOffice)"
                ],
                "summary": "Get Data Ticket",
                "operationId": "App\\Http\\Controllers\\Admin\\TicketController::scan",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Setup Ticket (BackOffice)"
                ],
                "summary": "Scan Ticket",
                "operationId": "App\\Http\\Controllers\\Admin\\TicketController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ticket_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/admin/ticket/hitory": {
            "get": {
                "tags": [
                    "TICKET"
                ],
                "summary": "TICKET HISTORY",
                "operationId": "App\\Http\\Controllers\\TicketController::history",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}